implemented basic ir parser

+++ ir/lib/value/constant_float.cpp
@@ -1,7 +1,7 @@
 #include <scc/ir/type.hpp>
 #include <scc/ir/value.hpp>
 
-scc::ir::ConstantFloat::ConstantFloat(FloatType::Ptr type, const double value)
+scc::ir::ConstantFloat::ConstantFloat(FloatType::Ptr type, const long double value)
     : Constant(std::move(type)),
       m_Value(value)
 {
@@ -12,7 +12,7 @@ std::ostream &scc::ir::ConstantFloat::PrintOperand(std::ostream &stream) const
     return m_Type->Print(stream) << " 0x" << std::hexfloat << m_Value << std::defaultfloat;
 }
 
-double scc::ir::ConstantFloat::GetValue() const
+long double scc::ir::ConstantFloat::GetValue() const
 {
     return m_Value;
 }
